Split and strip comment from lines
Why:

The comments sometimes end up falling within the range of the previous field.
This makes sure to strip them out.

This change addresses the need by:

Using .split()[0].strip()  for the last column.
It splits the expected last segment of the line, and takes the first part of it.
Anything beyond that point should be part of a comment.