Add standardize to mol2 output
Why:

The molecule_to_mol2 function uses a high-level writer,
which overwrites atomic properties such as atom names in an
effort to standardize them.
Sometimes it is desirable to keep the atom names provided in
the input.

This change addresses the need by:

* An option called `standardize` is added to molecule_to_mol2
* (Some) functions that depend on it have been updated to
make it accessible.
* Tests of the mol2 reading have been added to validate
that behavior is as expected.